change new() to self.new() for jruby compability. move extensions inside of externals to not polute namespace. Add lib to LOAD_PATH in Rakefile.
