+++ lib/externals/extensions/string.rb
