+++ lib/externals/extensions/symbol.rb
